+++ src/index.ts
@@ -1,9 +1,4 @@
-export { decode, decodeDocuments } from './cbor';
-export type {
-  DecodedDocuments,
-  IssuerSigned,
-  DocumentValue,
-  MDOC,
-} from './cbor/schema';
-export { sign, verify } from './cose';
-export type { COSESignResult } from './cose/types';
+import * as CBOR from './cbor';
+import * as COSE from './cose';
+
+export { CBOR, COSE };
